Move nodepool to this private repo
The nodepool.yaml.sample file is to be copied by the user and modified.
The script will manually copy this to the main repo where the template
will be filled in.

Also update the vars.sh to be a sample file & use the .gitignore
to facilitate keeping private files shared publicly.