Per convo on IRC with jeblair, removing gate
Removes the gate pipeline from the Zuul layout.yaml file. No external
systems are voting on the gate pipeline, so just remove it. Also removes
the precedence keyword from the check pipeline and the includes section.

asselin: Add back in the includes section which is needed for nodepool
single use nodes.

Conflicts:

	etc/zuul/layout.yaml